fix can not save record and some code issue
