--- database/activities.db
@@ -1,1 +1,0 @@
-{"location":"鼓岩國小","startTime":"18 : 00","endTime":"20 : 00","date":"2017-02-18","user":"5659012301974","_id":"kveefKnOGc6JTWa3"}
