+++ database/users.db
@@ -1,0 +1,1 @@
+{"userId":"U451048461b01b3d2b99c117dca0e306d","paymentRecord":[],"isPresident":false,"_id":"N64HNWSltGXf8b6m"}
